uniendo rama dev con main

+++ database/migrations/2023_05_05_200420_create_certificado_table.php
@@ -13,6 +13,21 @@ return new class extends Migration
     {
         Schema::create('certificados', function (Blueprint $table) {
             $table->id();
+            $table->string('tipo_solicitud');
+            $table->foreignId('modalidade_id')->constrained()->onDelete('cascade');
+            $table->string('no_control');
+            $table->string('especialidad');
+            $table->foreignId('turno_id')->constrained()->onDelete('cascade');
+            $table->string('nombre_alumno');
+            $table->string('curp');
+            $table->string('telefono');
+            $table->string('correo');
+            $table->string('folio_pago');
+            $table->boolean('bachillerato_doc');
+            $table->boolean('secundaria_doc');
+            $table->boolean('nacimiento_doc');
+            $table->boolean('curp_doc');
+            $table->boolean('pago_doc');
             $table->timestamps();
         });
     }
@@ -22,6 +37,28 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('certificados');
+        if (Schema::hasTable('certificados')) {
+            Schema::table('certificados', function (Blueprint $table) {
+                $table->dropForeign('certificados_modalidade_id_foreign');
+                $table->dropForeign('certificados_turno_id_foreign');
+                $table->dropColumn(
+                    [
+                        'tipo_solicitud', 
+                        'no_control', 
+                        'especialidad', 
+                        'nombre_alumno', 
+                        'curp', 
+                        'telefono', 
+                        'correo', 
+                        'folio_pago',
+                        'bachillerato_doc',
+                        'secundaria_doc',
+                        'nacimiento_doc',
+                        'curp_doc',
+                        'pago_doc'
+                    ]);
+            });
+            Schema::dropIfExists('certificados');
+        }
     }
 };
